DEV/Malahayu/malahayu/webapp/requirements.txt
Django==1.11.6
Pillow==5.0.0
django-wysiwyg-redactor==0.5.1